# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for NutriPlanner - Nutritionist and Patient Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients/{id}/diets:
    get:
      operationId: listPatientDiets
      tags: [diets]
      summary: List all diet plans for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of diet plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DietPlan"
  /diets:
    post:
      operationId: createDietPlan
      tags: [diets]
      summary: Create a new diet plan for a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDietPlanRequest"
      responses:
        "201":
          description: Diet plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DietPlan"
  /diets/{id}:
    get:
      operationId: getDietPlan
      tags: [diets]
      summary: Get a diet plan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Diet plan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DietPlan"
        "404":
          description: Diet plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDietPlan
      tags: [diets]
      summary: Update a diet plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDietPlanRequest"
      responses:
        "200":
          description: Diet plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DietPlan"
    delete:
      operationId: deleteDietPlan
      tags: [diets]
      summary: Delete a diet plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /diets/{id}/meals:
    get:
      operationId: listDietMeals
      tags: [meals]
      summary: List meals in a diet plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meal"
    post:
      operationId: addMealToDiet
      tags: [meals]
      summary: Add a meal to a diet plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealRequest"
      responses:
        "201":
          description: Meal added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
  /meals/{id}:
    put:
      operationId: updateMeal
      tags: [meals]
      summary: Update a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMealRequest"
      responses:
        "200":
          description: Meal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
    delete:
      operationId: deleteMeal
      tags: [meals]
      summary: Delete a meal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    DietPlan:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        totalCalories:
          type: number
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - patientId
        - name
        - isActive
        - createdAt
    CreateDietPlanRequest:
      type: object
      properties:
        patientId:
          type: integer
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
      required:
        - patientId
        - name
    UpdateDietPlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
    Meal:
      type: object
      properties:
        id:
          type: integer
        dietPlanId:
          type: integer
        name:
          type: string
        time:
          type: string
          description: Suggested time, e.g. "08:00"
        description:
          type: string
        foods:
          type: array
          items:
            $ref: "#/components/schemas/FoodItem"
        calories:
          type: number
        order:
          type: integer
      required:
        - id
        - dietPlanId
        - name
        - order
    CreateMealRequest:
      type: object
      properties:
        name:
          type: string
        time:
          type: string
        description:
          type: string
        foods:
          type: array
          items:
            $ref: "#/components/schemas/FoodItem"
        calories:
          type: number
        order:
          type: integer
      required:
        - name
        - order
    UpdateMealRequest:
      type: object
      properties:
        name:
          type: string
        time:
          type: string
        description:
          type: string
        foods:
          type: array
          items:
            $ref: "#/components/schemas/FoodItem"
        calories:
          type: number
        order:
          type: integer
    FoodItem:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
      required:
        - name
        - quantity
...