# lib/api-zod/src/generated/types/createDietPlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for NutriPlanner - Nutritionist and Patient Platform
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDietPlanRequest {
  patientId: number;
  name: string;
  description?: string;
  startDate?: Date;
  endDate?: Date;
  isActive?: boolean;
}